+++ src_james/solver_multimodel/SingleColorSolver.py
@@ -34,7 +34,7 @@ class SingleColorSolver(Solver):
 
     def action(self, grid, query=None, task=None):
         ratio  = task_shape_ratios(task)[0]
-        output = np.zeros(( int(grid.shape[0] * ratio[0]), int(grid.shape[1] * ratio[1]) ))
+        output = np.zeros(( int(grid.shape[0] * ratio[0]), int(grid.shape[1] * ratio[1]) ), dtype=np.int8)
         color  = query(grid) if callable(query) else query
         output[:,:] = color
         return output
